refactory directory structure
created more sub-directories and shifted logic for init.lua
down into jai/init.lua . This means I can have other configs
in the future and separate them out based on directories in
same level as 'jai'.
