+++ lua/jai/lsp/nvim_cmp_config.lua
