+++ lua/jai/settings/opts.lua
