+++ lua/jai/settings/vars.lua
@@ -36,7 +36,7 @@ g.t_co = 256
 g.background = "dark"
 
 -- update the packpath via adding additional search path
--- TODO: this is duplicated in jai.plugins.lua; requires refactoring
-local packer_path = vim.fn.stdpath('config') .. '/site'
-vim.o.packpath = vim.o.packpath .. ',' .. packer_path
+-- This is where neovim will search for packages
+local package_path = vim.fn.stdpath('config') .. '/site'
+vim.o.packpath = vim.o.packpath .. ',' .. package_path
